refactor(web): clean up connection-form naming conventions
Following form directory naming guidelines:
- Rename connectionFormFields.ts to fields.ts (directory context makes prefix redundant)
- Rename ConnectionForm.tsx to Form.tsx (main component in dedicated directory)
- Rename connectionFormOptions to defaultOptions (more generic, applicable to other forms)

All imports and references updated accordingly.
