+++ web/src/components/network/connection-form/BindingModeSelector.test.tsx
@@ -24,11 +24,11 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
-import { connectionFormOptions } from "./connectionFormFields";
+import { defaultOptions } from "./fields";
 import BindingModeSelector from "./BindingModeSelector";
 
 function TestForm() {
-  const form = useAppForm({ ...connectionFormOptions });
+  const form = useAppForm({ ...defaultOptions });
   return <BindingModeSelector form={form} />;
 }
