+++ web/src/components/network/connection-form/BondFields.test.tsx
@@ -24,7 +24,7 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { installerRender } from "~/test-utils";
 import { useAppForm } from "~/hooks/form";
-import { connectionFormOptions } from "./connectionFormFields";
+import { defaultOptions } from "./fields";
 import { DeviceState } from "~/types/network";
 import { CONNECTION_TYPE } from "~/utils/network";
 import BondFields from "./BondFields";
@@ -55,9 +55,9 @@ function TestForm({
   isEditing?: boolean;
 }) {
   const form = useAppForm({
-    ...connectionFormOptions,
+    ...defaultOptions,
     defaultValues: {
-      ...connectionFormOptions.defaultValues,
+      ...defaultOptions.defaultValues,
       name: "test-bond",
       type: CONNECTION_TYPE.BOND,
       ...defaultValues,
