+++ web/src/components/network/connection-form/Form.test.tsx
@@ -31,7 +31,7 @@ import {
   DeviceState,
 } from "~/types/network";
 import { CONNECTION_TYPE } from "~/utils/network";
-import ConnectionForm from "./ConnectionForm";
+import ConnectionForm from "./Form";
 
 const mockDevice1 = {
   name: "enp1s0",
